backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
google-generativeai>=0.3
python-dotenv>=1.0
gunicorn>=21.2
django-cors-headers